配置调试完成,提现调试了一半

+++ game_analyze/src/com/gamecenter/service/agent/AgentShopService.java
@@ -43,7 +43,7 @@ public interface AgentShopService extends BaseService<OpShop> {
 	 * @param type
 	 * @return 
 	 */  
-	OpShop findShopGoodsByPrice(int price,int type);
+	OpShop findShopGoodsByPrice(double price,int type);
 	
 }
